Refactor to split functions working with google sheets to its on module
